+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - Improve performance of prepare_traindatasets for large label files (#116)
 - Add check that train and validation data is mandatory (#118)
 - Support geopandas 0.13 + remove dependency on pygeos (#121)
+- Use [pygeoops](https://github.com/pygeoops/pygeoops) for inline simplification during
+  predict (#):
+   - 10x faster for inline simplification
+   - new LANG+ algorithm for improved simplification (= now default algorithm)
 - Several small improvements to logging, linting,... (#110, #112, #122,...)
 
 ### Bugs fixed